backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Backend written in FastAPI for interactive schools ranking on the map."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.127.0",
    "geoalchemy2>=0.18.1",
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.3.251219",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "sqlmodel>=0.0.28",
    "unidecode>=1.4.0",
]

[project.scripts]
data-import = "app.data_import.main:main"
data-transform = "app.data_import.transform:main"
data-scoring = "app.data_import.scoring:main"

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"  

[tool.uv.build-backend]
module-root = ""
module-name = ["app"]

[dependency-groups]
dev = [
    "basedpyright>=1.36.2",
    "debugpy>=1.8.19",
    "ruff>=0.14.10",
]

[tool.ruff]
exclude = [
    "**/__pycache__/**",
    "**/*.pyc",
    "venv/**",
    ".venv/**",
    "alembic/**",
]

[tool.ruff.lint]
select = [
    "E4",  # pycodestyle (specifically import conventions like E402)
    "E7",  # pycodestyle errors (comparisons, bare excepts)
    "E9",  # pycodestyle runtime/syntax errors 
    "F",   # Pyflakes (undefined names, unused imports/variables, etc.)
    "B",   # flake8-bugbear (potential bugs and design problems)
    "I",   # isort (import sorting)
    "Q",   # flake8-quotes (quote consistency)
    "ARG", # flake8-unused-arguments (detect unused function arguments)
    "FAST",# flake8-fastapi (FastAPI specific linting rules)
    "N",   # pep8-naming (PEP 8 naming conventions)
    "RUF"  # Ruff-specific rules (built-in rules unique to Ruff)
]

[tool.basedpyright]
exclude = [
    "**/__pycache__/**",
    "**/*.pyc",
    "venv/**",
    ".venv/**",
    "alembic/**",
]
# using ruff for unused code (ARG rule)
reportUnusedVariable = "none"
reportUnusedImport = "none"
reportUnusedFunction = "none"
reportUnusedClass = "none"
reportUnusedExpression = "none"
reportImportCycles = "none" # using ORMs it is almost impossible to avoid cycles
